pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "f1-news-pipeline"
version = "0.1.0"
description = "Automated F1 news discovery, scraping, and summarization pipeline"
requires-python = ">=3.9"

dependencies = [
    "apache-airflow[postgres]",
    "beautifulsoup4",
    "feedparser",
    "hdbscan",
    "mistralai",
    "protobuf",
    "python-dotenv",
    "python-dateutil",
    "requests",
    "scikit-learn",
    "SQLAlchemy",
    "alembic",
    "psycopg[binary]",
]

[tool.setuptools.packages.find]
where = ["."]
